Fixing timestep issue for refined meshes.
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@1912 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/had_amr/amr_c/stencil_data.hpp
@@ -17,7 +17,7 @@ struct stencil_data
 {
     size_t max_index_;   // overall number of data points
     size_t index_;       // sequential number of this data point (0 <= index_ < max_values_)
-    size_t timestep_;    // current time step
+    double timestep_;    // current time step
     size_t level_;    // refinement level
     double value_;            // current value
     double x_;      // x coordinate value
